requirements.txt
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.13.1
fsspec==2024.2.0
idna==3.8
imageio==2.35.1
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.2.1
numpy==1.26.3
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.2.0
pip==24.2
pipreqs==0.5.0
platformdirs==4.3.2
prompt_toolkit==3.0.47
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
pywin32==306
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==72.1.0
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
sympy==1.12
tdqm==0.0.1
threadpoolctl==3.5.0
tifffile==2024.8.30
tinycss2==1.3.0
torch==2.4.1+cu124
torchaudio==2.4.1+cu124
torchvision==0.19.1+cu124
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.9.0
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.43.0
yarg==0.1.9

pandas~=2.2.3
opencv-python~=4.11.0.86
torchmetrics~=1.5.1
pytorch-lightning~=2.4.0